// api/src/modules/database/constants/converter-names.ts
export enum ConverterNames {
  ACCESS_CARD = 'DATABASE/ACCESS_CARD_CONVERTER',
  AGENCY = 'DATABASE/AGENCY_CONVERTER',
  CLIENT = 'DATABASE/CLIENT_CONVERTER',
  DOER = 'DATABASE/DOER_CONVERTER',
  ENTRY_POINT = 'DATABASE/ENTRY_POINT_CONVERTER',
  JOB_BOARD = 'DATABASE/JOB_BOARD_CONVERTER',
  MISSION = 'DATABASE/MISSION_CONVERTER',
  QUESTION = 'DATABASE/QUESTION_CONVERTER',
  QUESTION_MESSAGE = 'DATABASE/QUESTION_MESSAGE_CONVERTER',
  QUESTION_SHEET = 'DATABASE/QUESTION_SHEET_CONVERTER',
  QUIZZ = 'DATABASE/QUIZZ_CONVERTER',
  QUIZZ_SHEET = 'DATABASE/QUIZZ_SHEET_CONVERTER',
}