// api/src/modules/database/constants/repository-names.ts
export enum RepositoryNames {
  ACCESS_CARD = 'DATABASE/ACCESS_CARD_REPOSITORY',
  AGENCY = 'DATABASE/AGENCY_REPOSITORY',
  CLIENT = 'DATABASE/CLIENT_REPOSITORY',
  DOER = 'DATABASE/DOER_REPOSITORY',
  ENTRY_POINT = 'DATABASE/ENTRY_POINT_REPOSITORY',
  JOB_BOARD = 'DATABASE/JOB_BOARD_REPOSITORY',
  MISSION = 'DATABASE/MISSION_REPOSITORY',
  QUESTION = 'DATABASE/QUESTION_REPOSITORY',
  QUESTION_SHEET = 'DATABASE/QUESTION_SHEET_REPOSITORY',
  QUIZZ = 'DATABASE/QUIZZ_REPOSITORY',
  QUIZZ_SHEET = 'DATABASE/QUIZZ_SHEET_REPOSITORY',
}